refactor: decouple Compose from social publishing, make Socials tab platform-agnostic
Compose tab is now blog/wire only — no social platform targets, no
caption derivation from Gutenberg content, no social meta writing.

Socials tab replaces the hardcoded Instagram pane with a generic
PlatformPublishPane that works for any authenticated DM Socials
platform. Platforms are listed with click-to-select, and each gets
its own caption textarea, image management, publish/review actions,
and comment management via the generic comments API.

Removed: deriveCaption(), deriveImageUrls(), social target loading
from Compose, entire instagram/ directory.

Added: socials/publish/index.ts — platform-agnostic publishing pane.
Updated: CSS for clickable/active platform list items.
