+++ src/blocks/studio/tabs/socials/index.ts
@@ -6,7 +6,7 @@ import type { SocialPlatformsResponse } from '@extrachill/api-client';
 
 import { studioClient } from '../../app/client';
 import type { StudioPaneProps } from '../../types/studio';
-import InstagramPane from './instagram';
+import PlatformPublishPane from './publish';
 import type { SocialPlatformConfig } from '../../types/externals';
 
 const h = createElement as typeof import( 'react' ).createElement;
@@ -20,12 +20,14 @@ interface PlatformEntry {
 	authenticated: boolean;
 	username: string | null;
 	type: string;
+	config: SocialPlatformConfig;
 }
 
 const SocialsPane = ( _props: StudioPaneProps ): ReactElement | null => {
 	const [ platforms, setPlatforms ] = useState< SocialPlatformsResponse >( {} );
 	const [ error, setError ] = useState( '' );
 	const [ isLoading, setIsLoading ] = useState( true );
+	const [ activePlatform, setActivePlatform ] = useState< string | null >( null );
 
 	useEffect( () => {
 		const loadPlatforms = async (): Promise< void > => {
@@ -50,25 +52,33 @@ const SocialsPane = ( _props: StudioPaneProps ): ReactElement | null => {
 		const platformConfig = config as SocialPlatformConfig | undefined;
 
 		return {
-		slug,
-		label: platformConfig?.label || slug,
-		authenticated: platformConfig?.authenticated || false,
-		username: platformConfig?.username || null,
-		type: platformConfig?.type || 'publish',
+			slug,
+			label: platformConfig?.label || slug,
+			authenticated: platformConfig?.authenticated || false,
+			username: platformConfig?.username || null,
+			type: platformConfig?.type || 'publish',
+			config: platformConfig || { label: slug },
 		};
 	} );
 
 	const connectedPlatforms = availablePlatforms.filter( ( item ) => item.authenticated );
-	const publishPlatforms = availablePlatforms.filter( ( item ) => item.type !== 'fetch' );
+	const publishablePlatforms = connectedPlatforms.filter( ( item ) => item.type !== 'fetch' );
+
+	// Auto-select first publishable platform if none is active.
+	useEffect( () => {
+		if ( ! activePlatform && publishablePlatforms.length > 0 ) {
+			setActivePlatform( publishablePlatforms[ 0 ].slug );
+		}
+	}, [ activePlatform, publishablePlatforms.length ] );
 
 	if ( isLoading ) {
 		return h(
 			'div',
 			{ className: 'ec-studio-pane ec-studio-pane--socials' },
-				h(
-					PanelView,
-					{ className: 'ec-studio-panel', compact: true },
-					h( InlineStatusView, { tone: 'info', className: 'ec-studio-message' }, __( 'Loading social platforms…', 'extrachill-studio' ) )
+			h(
+				PanelView,
+				{ className: 'ec-studio-panel', compact: true },
+				h( InlineStatusView, { tone: 'info', className: 'ec-studio-message' }, __( 'Loading social platforms…', 'extrachill-studio' ) )
 			)
 		);
 	}
@@ -77,13 +87,15 @@ const SocialsPane = ( _props: StudioPaneProps ): ReactElement | null => {
 		return h(
 			'div',
 			{ className: 'ec-studio-pane ec-studio-pane--socials' },
-				h(
-					PanelView,
-					{ className: 'ec-studio-panel', compact: true },
-					h( InlineStatusView, { tone: 'error', className: 'ec-studio-message' }, error )
+			h(
+				PanelView,
+				{ className: 'ec-studio-panel', compact: true },
+				h( InlineStatusView, { tone: 'error', className: 'ec-studio-message' }, error )
 			)
 		);
 	}
+
+	const selectedPlatform = publishablePlatforms.find( ( p ) => p.slug === activePlatform ) || null;
 
 	return h(
 		'div',
@@ -91,16 +103,34 @@ const SocialsPane = ( _props: StudioPaneProps ): ReactElement | null => {
 		h(
 			'div',
 			{ className: 'ec-studio-pane__grid' },
-				h(
-					PanelView,
+			h(
+				PanelView,
 				{ className: 'ec-studio-panel', compact: true },
-				createElement( 'p', null, sprintf( __( '%d connected, %d publish-capable.', 'extrachill-studio' ), connectedPlatforms.length, publishPlatforms.length ) ),
+				h( PanelHeader, {
+					description: sprintf(
+						__( '%d platform(s) connected, %d publish-capable. Select a platform below to compose and publish.', 'extrachill-studio' ),
+						connectedPlatforms.length,
+						publishablePlatforms.length
+					),
+				} ),
 				createElement(
 					'ul',
 					{ className: 'ec-studio-social-platforms' },
 					...availablePlatforms.map( ( item ) => createElement(
 						'li',
-						{ key: item.slug, className: 'ec-studio-social-platforms__item' },
+						{
+							key: item.slug,
+							className: [
+								'ec-studio-social-platforms__item',
+								item.authenticated && item.type !== 'fetch' ? 'ec-studio-social-platforms__item--clickable' : '',
+								activePlatform === item.slug ? 'ec-studio-social-platforms__item--active' : '',
+							].filter( Boolean ).join( ' ' ),
+							onClick: item.authenticated && item.type !== 'fetch'
+								? () => setActivePlatform( item.slug )
+								: undefined,
+							role: item.authenticated && item.type !== 'fetch' ? 'button' : undefined,
+							tabIndex: item.authenticated && item.type !== 'fetch' ? 0 : undefined,
+						},
 						createElement(
 							'span',
 							{ className: 'ec-studio-social-platforms__name' },
@@ -123,19 +153,24 @@ const SocialsPane = ( _props: StudioPaneProps ): ReactElement | null => {
 					) )
 				)
 			),
-				h(
-					PanelView,
+			h(
+				PanelView,
 				{ className: 'ec-studio-panel', compact: true },
-				createElement( 'p', null, __( 'Each connected platform has its own publishing workflow below. New platforms appear automatically when connected in Data Machine.', 'extrachill-studio' ) ),
-				createElement(
-					'ul',
-					null,
-					createElement( 'li', null, sprintf( __( '%d of %d platform(s) connected.', 'extrachill-studio' ), connectedPlatforms.length, availablePlatforms.length ) ),
-					createElement( 'li', null, __( 'Instagram — publish posts, manage comments, submit drafts for review.', 'extrachill-studio' ) )
-				)
+				createElement( 'p', null, __( 'Connected platforms appear automatically when authenticated in Data Machine. Select a platform from the list to compose and publish content.', 'extrachill-studio' ) ),
+				publishablePlatforms.length === 0
+					? h( InlineStatusView, { tone: 'warning', className: 'ec-studio-message' }, __( 'No publish-capable platforms are connected yet. Authenticate a platform in Data Machine Socials to get started.', 'extrachill-studio' ) )
+					: null
 			)
 		),
-		h( InstagramPane )
+		selectedPlatform
+			? h( PlatformPublishPane, {
+				key: selectedPlatform.slug,
+				slug: selectedPlatform.slug,
+				label: selectedPlatform.label,
+				username: selectedPlatform.username,
+				config: selectedPlatform.config,
+			} )
+			: null
 	);
 };
